# nextjs/public/openapi.yaml
openapi: 3.1.0
info:
  title: VidTempla API
  version: "1.0"
  description: |
    VidTempla is an API-first platform for managing YouTube video descriptions with templates, containers, and variables.
    The REST API lets AI agents securely manage YouTube channels without handling OAuth directly.

    ## Authentication
    All requests require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer vtk_your_key_here
    ```
    API keys have one of two permission levels:
    - **read** — can only fetch data (GET requests and POST /analytics/query)
    - **read-write** — can also create, update, and delete resources

    Generate keys from the dashboard at Settings > API Keys.

    ## Response Envelope
    Every response uses a standard envelope:
    ```json
    { "data": { ... }, "error": null, "meta": { ... } }
    ```

    ## Pagination
    All list endpoints use cursor-based pagination. Default limit is 50, max is 100.
    ```
    GET /videos?cursor=...&limit=50
    ```
  contact:
    name: VidTempla
    url: https://vidtempla.com
  license:
    name: Proprietary
servers:
  - url: https://vidtempla.com/api/v1
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key starting with vtk_. Generate from Settings > API Keys."
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          type: "null"
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            suggestion:
              type: string
              description: Actionable suggestion for agents to self-correct
            status:
              type: integer
              description: HTTP status code
          required: [code, message, status]
      required: [data, error]
...
    SuccessMessage:
      type: object
      properties:
        data:
          type: object
          properties:
            success:
              type: boolean
              example: true
        error:
          type: "null"
...
    Video:
      type: object
      properties:
        id:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        videoId:
          type: string
          description: YouTube video ID
        title:
          type: string
          nullable: true
        currentDescription:
          type: string
          nullable: true
        containerId:
          type: string
          format: uuid
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        channel:
          type: object
          properties:
            id:
              type: string
              format: uuid
            channelId:
              type: string
            title:
              type: string
              nullable: true
        container:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
...
    RetentionPoint:
      type: object
      description: A single point on the 100-point audience retention curve from YouTube Analytics (elapsedVideoTimeRatio report).
      properties:
        position:
          type: number
          description: "Elapsed video time ratio (0.01 to 1.0 in 0.01 increments). Each step represents 1% of total video duration. For example, 0.50 = the halfway point of the video."
          minimum: 0.01
          maximum: 1.0
          example: 0.50
        watchRatio:
          type: number
          description: "Absolute watch ratio — the fraction of viewers still watching at this position (0.0 to 1.0). A value of 0.75 means 75% of viewers who started the video are still watching at this point."
          minimum: 0.0
          maximum: 1.0
          example: 0.75
        relativePerformance:
          type: number
          description: "How this video's retention compares to YouTube videos of similar length. 1.0 = average, >1.0 = above average, <1.0 = below average."
          example: 1.12
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INVALID_API_KEY
              message: API key not found or has been revoked
              suggestion: Generate a new key from Settings > API Keys
              status: 401
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
              suggestion: Try again later
              status: 500
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /videos/{id}:
    get:
      operationId: getVideoDetails
      summary: Get video details
      description: Returns VidTempla metadata for a video, enriched with real-time YouTube statistics when available.
      tags: [Videos]
      x-quota-cost: 0
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: VidTempla UUID or YouTube video ID (e.g. `dQw4w9WgXcQ`)
          examples:
            uuid:
              summary: VidTempla internal UUID
              value: "296bbf0c-2fc2-47bc-8ab9-a2072245528a"
            youtubeId:
              summary: YouTube video ID
              value: "dQw4w9WgXcQ"
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Video"
                      - type: object
                        properties:
                          youtube:
                            type: object
                            nullable: true
                            description: Real-time YouTube video resource (snippet, statistics, contentDetails)
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /videos/{id}/analytics:
    get:
      operationId: getVideoAnalytics
      summary: Get video analytics
      description: Fetches analytics data from the YouTube Analytics API for a specific video. Defaults to last 28 days.
      tags: [Videos]
      x-quota-cost: 0
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: VidTempla UUID or YouTube video ID (e.g. `dQw4w9WgXcQ`)
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Start date (YYYY-MM-DD). Defaults to 28 days ago.
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: End date (YYYY-MM-DD). Defaults to today.
        - name: metrics
          in: query
          schema:
            type: string
            default: "views,estimatedMinutesWatched,averageViewDuration"
          description: Comma-separated YouTube Analytics metrics
        - name: dimensions
          in: query
          schema:
            type: string
            default: day
          description: YouTube Analytics dimension
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: YouTube Analytics API response
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /videos/{id}/retention:
    get:
      operationId: getVideoRetention
      summary: Get audience retention data
      description: |
        Returns a 100-point audience retention curve from the YouTube Analytics API.
        Each point represents 1% of the video's duration (position 0.01 = 1%, position 1.0 = 100%).
        Useful for understanding where viewers drop off and identifying engaging or weak segments.
      tags: [Videos]
      x-quota-cost: 0
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: VidTempla UUID or YouTube video ID (e.g. `dQw4w9WgXcQ`)
      responses:
        "200":
          description: Retention curve
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RetentionPoint"
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /videos/{id}/assign:
    post:
      operationId: assignVideoToContainer
      summary: Assign video to a container
      description: Assigns an unassigned video to a container and initializes template variables. Returns 400 if the video is already assigned.
      tags: [Videos]
      x-quota-cost: 0
      x-requires-write: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: VidTempla UUID or YouTube video ID (e.g. `dQw4w9WgXcQ`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [containerId]
              properties:
                containerId:
                  type: string
                  format: uuid
                  description: The container to assign the video to
            example:
              containerId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        "200":
          description: Video assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: Missing containerId or video already assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Insufficient permissions or video limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...